refactor(VirtualizedTable): separate cell renders and row classnames getter to utils
